develop-requirements.txt
apache-airflow==2.2.5
pytest==6.2.5
black==21.12b0
flake8==4.0.1
psycopg2==2.9.3
pytest-env==0.6.2
tzlocal==4.1
jmespath==1.0.0